requirements.txt
streamlit
torch
torchvision
Pillow
matplotlib
transformers
timm
numpy
pandas
tqdm